Implemented the missing signal valueChanged(void) in a wrapper for EACH control. Before there were just wrappers for doubleslider and checkbox, now there are wrappers for all used widget-control-types.

+++ src/gui/mainwindow.h
@@ -43,7 +43,7 @@ namespace gui
       void loadShaderDialog();
       void loadImageFile();
       void saveImage();
-      void showControls( unsigned short int index );
+      void showControls( int index );
 
   signals:
       void exitProgram();
